+++ include/cone_light_input_handler.h
@@ -19,12 +19,12 @@ private:
                  m_fuzzy_btn_c_state = BUTTON_RELEASED,
                  m_fuzzy_lid_state = LID_OPENED;
 
-  long m_btn_a_state_changed_ms = 0,
-       m_btn_b_state_changed_ms = 0,
-       m_btn_c_state_changed_ms = 0,
-       m_lid_state_changed_ms = 0;
+  uint32_t m_btn_a_state_changed_ms = 0,
+           m_btn_b_state_changed_ms = 0,
+           m_btn_c_state_changed_ms = 0,
+           m_lid_state_changed_ms = 0;
 
-  void handle_button(ConeLightButton btn, int raw_state, ConeLightEvent &fuzzy_state, ConeLightEvent &previous_state, long &state_changed_ms);
+  void handle_button(ConeLightButton btn, int raw_state, ConeLightEvent &fuzzy_state, ConeLightEvent &previous_state, uint32_t &state_changed_ms);
   void handle_lid(int raw_state);
 
 public:
